Regenerate Nix setup based on latest upstream docs
I started hitting various weird errors, recompiling GHC etc. so it seemed a good time to switch to the more modern approach.

Refactoring has been performed in order to reduce the diff about as much as possible. In particular, Hix seems to be designed to auto-pickup everything in `final.haskell-nix.hix.project` config from an external `hix.nix` file, but we've kept it all in one file instead.

We've taken some inessential suggestions from the template: including more cross-compilation targets, and adding dev tools to the shell. On the other hand we've disregarded the suggestion to include `legacyPackages`, which doesn't make any noticeable difference.